% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{edg2idx}
\alias{edg2idx}
\title{convert edge table to named parent vector}
\usage{
edg2idx(edges, parentCol = 1, childCol = 2)
}
\arguments{
\item{edges}{table with parent and child columns, each row
is an edge in a tree}

\item{parentCol}{column with parents}

\item{childCol}{column with children}
}
\description{
convert edge table to named parent vector
}
