% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get.children}
\alias{get.children}
\title{simple children retrieval}
\usage{
get.children(id, tax, skip = NULL)
}
\arguments{
\item{id}{a taxon ID}

\item{tax}{NCBI taxonomy object}

\item{skip}{vectors of ids to "skip", if the node 'id' is already
present in \code{skip}, the recursive function is not called}
}
\description{
returns a vector of all children
}
\examples{
get.children("1148", tax)
}
