% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get.parents}
\alias{get.parents}
\title{simple lineage retrieval}
\usage{
get.parents(id, tax, skip = NULL)
}
\arguments{
\item{id}{a taxon ID}

\item{tax}{NCBI taxonomy object}

\item{skip}{edges table to skip: if the node \code{id} is already present,
the recursive function is not called}
}
\description{
returns a simple lineage vector
}
\examples{
get.parents("1148", tax)
get.parents("1140", tax)
}
