% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getAllParents}
\alias{getAllParents}
\title{get common parents for a list of taxon IDs}
\usage{
getAllParents(ids, tax, root = "1", names = FALSE, dbug = FALSE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{root}{taxon ID of the taxonomic class to be considered as root}

\item{names}{return taxon names instead of ids}

\item{dbug}{print debug messages}
}
\description{
Loops over getParents for severals taxon IDs; avoids multiple
lineage retrieval by passing existing edges to the recursive function.
}
\examples{
getAllParents(c("1148","1140"), tax)
}
