% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getAncestor}
\alias{getAncestor}
\title{get lowest common ancestor}
\usage{
getAncestor(ids, tax, root = "1", names = FALSE, dbug = FALSE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{root}{taxon ID of the taxonomic class to be considered as root}

\item{names}{return taxon names instead of ids}

\item{dbug}{print debug messages}
}
\description{
retrieves the lowest common ancestor
}
\examples{
getAncestor(ids=c("1148","1140"), tax)
}
