% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getChildren}
\alias{getChildren}
\title{get all chilrden of a node}
\usage{
getChildren(id, tax, names = FALSE)
}
\arguments{
\item{id}{a single taxon ID}

\item{tax}{NCBI taxonomy object}

\item{names}{return taxon names instead of ids}
}
\description{
Returns edge table of all descendents. Performs
a depth-first search thus returns ordered edges
(order "cladewise" in package ape).
}
