% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getID}
\alias{getID}
\title{gets taxon IDs from taxon names}
\usage{
getID(names, tax)
}
\arguments{
\item{names}{taxonomy names}

\item{tax}{NCBI taxonomy object}
}
\description{
works for exact names (in ncbi file names.dmp), use
\code{\link{grepName}} to find IDs from
search patterns/partial names
}
\examples{
getID(c("Firmicutes","Cyanobacteria"), tax)
}
