% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getLineage}
\alias{getLineage}
\title{get NCBI taxonomy rank lineages}
\usage{
getLineage(ids, tax, ranks = c("phylum", "species"), names = FALSE,
  reduce = FALSE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{ranks}{vector of ranks to retrieve}

\item{names}{return taxon names instead of ids}

\item{reduce}{call \code{\link{reduceTaxonomy}} before search; untested
but this makes subsequent search faster by using \code{\link{getAllParents}}
first to reduce taxonomy to sub-tree for \code{ids}}
}
\description{
Returns a table with taxonomic lineages at the requested
rank levels. A wrapper around \code{\link{getAllParents}}.
}
\examples{
getLineage(c("1148","1140"), tax, names=TRUE,
         ranks=c("superkingdom","phylum","species"))
}
