% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getName}
\alias{getName}
\title{gets taxon names from taxon IDs}
\usage{
getName(ids, tax)
}
\arguments{
\item{ids}{vector or edge table of taxon IDs}

\item{tax}{NCBI taxonomy object}
}
\description{
gets taxon names from taxon IDs
}
\examples{
getName(c("1148","1140"), tax)
}
