% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getNames}
\alias{getNames}
\title{gets taxon names from taxon IDs}
\usage{
getNames(ids, tax)
}
\arguments{
\item{ids}{vector or edge table of taxon IDs}

\item{tax}{NCBI taxonomy object}
}
\description{
gets taxon names from taxon IDs
}
