% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getRanks}
\alias{getRanks}
\title{gets taxon ranks from taxon IDs}
\usage{
getRanks(ids, tax)
}
\arguments{
\item{ids}{vector or edge table of taxon IDs}

\item{tax}{NCBI taxonomy object}
}
\description{
gets taxon ranks from taxon IDs
}
\examples{
getRanks(c("1148","1140"), tax)
}
