% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{grepName}
\alias{grepName}
\title{find taxon ID for species names}
\usage{
grepName(pattern, tax, ...)
}
\arguments{
\item{pattern}{regular expression, see argument \code{pattern}
in \code{\link{grep}}}

\item{tax}{NCBI taxonomy object}

\item{...}{arguments to \code{\link{grep}}}
}
\description{
Simply uses base R's \code{\link{grep}} to grep for a
name pattern in the official NCBI taxon names and returns
all matching taxon IDs as a table with columns ID and name.
}
\examples{
grepName("Synechocystis", tax)
}
