% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{idx2edg}
\alias{idx2edg}
\title{convert named parent vector to edge table}
\usage{
idx2edg(idx)
}
\arguments{
\item{idx}{vector of parents, named by children}
}
\description{
NOTE: only for trees, does not work for cyclic graphs or DAG.
}
