% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{package}
\name{ncbitax}
\alias{ncbitax}
\alias{ncbitax-package}
\title{ncbitax: a simple NCBI taxonomy parser}
\description{
Similar to functions provided by ape and phylobase,
but specific for NCBI Taxonomy.
}
\author{
Rainer Machne \email{machne@hhu.de}
}
