% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{parseNCBITaxonomy}
\alias{parseNCBITaxonomy}
\title{load tbi taxonomy files}
\usage{
parseNCBITaxonomy(taxd, names = TRUE, ranks = TRUE, merged = TRUE,
  verb = TRUE)
}
\arguments{
\item{taxd}{path to unpacked NCBI taxonomy (taxdump.tar.gz)}

\item{names}{load species names, takes long to load!}

\item{ranks}{load rank names}

\item{merged}{load list of merged (outdated) taxon IDs and their replacement}

\item{verb}{print progress messages}
}
\description{
load tbi taxonomy files
}
\examples{
taxdb <- "/data/taxonomy/db"
tax <- parseNCBITaxonomy(taxdb)
}
