% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{reduceTaxonomy}
\alias{reduceTaxonomy}
\title{reduced taxonomy tree for faster searches}
\usage{
reduceTaxonomy(ids, tax)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}
}
\description{
This reduces the full NCBI taxonomy to the subset
containing all taxons in \code{ids}. This makes all
subsequent searches much faster.
}
\examples{
rtax <- reduceTaxonomy(c("1148","1140"), tax)
lapply(tax, length)
lapply(rtax, length)
}
