% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax2newick}
\alias{tax2newick}
\title{convert NCBI taxonomy sub trees to newick format}
\usage{
tax2newick(ids, tax, names = FALSE, full = FALSE, ranks = FALSE,
  verb = FALSE, dbug = FALSE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{names}{replace taxon IDs with names}

\item{full}{if  \code{TRUE} \code{singleton} nodes with only
one child will NOT be collapsed}

\item{ranks}{add rank information for nodes}

\item{verb}{print progress messages}

\item{dbug}{print debug messages}
}
\description{
convert NCBI taxonomy sub trees to newick format
}
\examples{
new <- tax2newick(ids=c("1148","1140"), tax=tax, names=TRUE)
}
