% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax2phylo}
\alias{tax2phylo}
\title{converts an NCBI taxonomy object to an \code{ape::phylo} tree}
\usage{
tax2phylo(tax, root = "1", names = FALSE, order = FALSE)
}
\arguments{
\item{tax}{NCBI taxonomy object}

\item{root}{taxon ID of the taxonomic class to be considered as root}

\item{names}{replace taxon IDs by names}

\item{order}{order edges, required for attribute "cladewise" of
the returned \code{ape::phylo} tree}
}
\description{
converts an NCBI taxonomy object to an \code{ape::phylo} tree
}
\examples{
phy <- tax2phylo(tax, root="1890428", order=TRUE, names=TRUE)
ape::plot.phylo(phy)
}
