% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{updateIDs}
\alias{updateIDs}
\title{update a list of taxonomy IDs}
\usage{
updateIDs(ids, tax, verb = TRUE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{verb}{print progress messages}
}
\description{
Looks in merged node list from file \code{merged.dmp}, whether
the passed \code{ids} have been replaced (merged) and returns
a list of updated (where in merged) ids.
It is recommended to run this on all IDs before using other
functions of this package!
}
