% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getParents}
\alias{getParents}
\title{lineage retrieval}
\usage{
getParents(id, tax, skip = NULL, root = "1", names = FALSE,
  verb = TRUE, dbug = FALSE)
}
\arguments{
\item{id}{a single taxon ID}

\item{tax}{NCBI taxonomy object}

\item{skip}{edges table to skip, if the node \code{id} is already present,
the recursive function is not called (used internally
from \code{\link{getAllParents}})}

\item{root}{taxon ID of the taxonomic class to be considered as root}

\item{names}{return taxon names instead of ids}

\item{verb}{print progress messages}

\item{dbug}{print debug messages}
}
\description{
Returns an edge table of the parent lineage until \code{root}.
}
\examples{
getParents("1140", tax)
getParents("1148", tax, root="1783272")
getParents("1048", tax, root="1783272") # root not in lineage!
}
