% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{getRank}
\alias{getRank}
\title{return requested taxonomy ranks, starting from taxid}
\usage{
getRank(ids, tax, ranks = c("phylum", "species"), names = FALSE)
}
\arguments{
\item{ids}{vector of taxon IDs}

\item{tax}{NCBI taxonomy object}

\item{ranks}{vector of ranks to retrieve}

\item{names}{return taxon names instead of ids}
}
\description{
return requested taxonomy ranks, starting from taxid
}
\examples{
getRank(c("1148","1140"), tax, names=TRUE,
        ranks=c("superkingdom","phylum","species"))
}
